varios

+++ TurismoRealEscritorio/Modelos/Departamento.cs
@@ -10,7 +10,7 @@ namespace TurismoRealEscritorio.Modelos
         public int Arriendo { get; set; }
         public int Habitaciones { get; set; }
         public int Banos { get; set; }
-        public int Mts_cuadrados { get; set; }
+        public float Mts_cuadrados { get; set; }
         public int Contribuciones { get; set; }
         public int Dividendo { get; set; }
         public int Id_localidad { get; set; }
